refactor(FaceCapture): 웹캠 로직과 UI 분리
